// src/types/LevelReadyPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProxyLevel } from "./ProxyLevel";

export type LevelReadyPayload = { imageId: string, level: ProxyLevel, rev: number, width: number, height: number, flip: number, hasColorProfile: boolean | null, colorMatrix: Array<number> | null, };